Adding forced contact query/pull in policy query method. This is a side effect of going to a pull-only strategy for policies (i.e. no pushing from the plugin).
This was added to handle cases when a policy comes in scope for a user, but the contact for that policy would otherwise be out of scope (as in it has no in-scope opportunities). In these cases, the contact would never make it to the phone (because the plugin will never push it) and therefore policies meeting this criteria would never be shown to the user.

Also repointed deployment scripts at leaders_meeting branch for model.