Added new push_objects utility that will only commit an update to redis if the object being updated already exists.
This helps to mitigate scenarios where a client-initiated update causes an opp to go out of scope (i.e. opp marked as lost from phone), which then causes the plugin to push a delete to RhoSync. In some cases this delete would get pushed to rhosync before the update method in the source adapter was finished with its update from the client, which would cause the push_objects call after the proxy update succeeds to create a partial opportunity object in redis.