Implemented more robust form of conflict management; now uses a new field managed from the client side to track when client updates were made in UTC.
Old method relied on the last activity date field from the client, which did not have timezone information. This resulted in issues if the phone was in a timezone other than central, as conflict management would either produce false positives if the user was in a time zone west of central or false negatives if the user was in a time zone east of central.